validacion rutas nivel educacion

+++ controllers/NivelEducacionController.js
@@ -1,4 +1,5 @@
 import NivelEducacion from "../models/NivelEducacion.js";
+import Jornada from "../models/Jornada.js";
 import { decodeJWT } from "../utils/codificar.js";
 
 export const getAllNivelesEducacion = async (req, res) => {
@@ -36,6 +37,16 @@ export const getNivelEducacion = async (req, res) => {
 
 export const createNivelEducacion = async (req, res) => {
     try {
+        const token = req.cookies?.token;
+        const userData = decodeJWT(token);
+
+        if (!userData) {
+            return res.status(403).json({ status: 'ERROR', message: "Token inválido o no proporcionado" });
+        }
+
+        req.body.id_colegio = userData.id_colegio
+        req.body.id_usuario_creo = userData.id
+
         await NivelEducacion.create(req.body);
         res.status(200).json({ status: 'OK', message: 'Nivel de educación creado correctamente!' });
     } catch (error) {
@@ -45,6 +56,15 @@ export const createNivelEducacion = async (req, res) => {
 
 export const updateNivelEducacion = async (req, res) => {
     try {
+        const token = req.cookies?.token;
+        const userData = decodeJWT(token);
+
+        if (!userData) {
+            return res.status(403).json({ status: 'ERROR', message: "Token inválido o no proporcionado" });
+        }
+
+        req.body.id_usuario_modifico = userData.id
+
         await NivelEducacion.update(req.body, {
             where: {
                 id_nivel: req.params.id
@@ -62,6 +82,12 @@ export const deleteNivelEducacion = async (req, res) => {
 
         if (!nivelEducacion) {
             return res.status(400).json({ status: 'ERROR', message: 'Nivel de educación no encontrado' });
+        }
+
+        const jornada = await Jornada.findOne({ where: { id_nivel: req.params.id }})
+
+        if (jornada) {
+            return res.status(400).json({ status: 'ERROR', message: 'Nivel de Educación asociado a una Jornada' });
         }
 
         await NivelEducacion.destroy({
